+++ src/modules/common_manipulators/term_evaluator.f90
@@ -29,10 +29,13 @@ module term_evaluator_class
 
     type ,public ,extends(Manipulator) :: TermEvaluator
         !! Manipulator that evaluates a specific model,term pairs and writes the total result into variable with given index
+        !! Optionally accumulates the values in the variable instead of overriding them.
 
         integer(ik)                                  ,private :: resultVarIndex 
         integer(ik)       ,allocatable ,dimension(:) ,private :: evaluatedModelIndex 
         type(StringArray) ,allocatable ,dimension(:) ,private :: evaluatedTermName
+
+        logical ,private :: accumulate = .false. !! If true will accumulate values instead of overriding them
 
         contains
 
@@ -44,13 +47,14 @@ module term_evaluator_class
     interface
 !-----------------------------------------------------------------------------------------------------------------------------------
 !-----------------------------------------------------------------------------------------------------------------------------------
-    pure module subroutine initEvaluator(this,resultVarIndex,modelIndices,termNames) 
+    pure module subroutine initEvaluator(this,resultVarIndex,modelIndices,termNames,accumulate) 
         !! TermEvaluator initialization routine
 
         class(TermEvaluator)            ,intent(inout)  :: this
         integer(ik)                     ,intent(in)     :: resultVarIndex !! Index of variable to write the result in
         integer(ik)       ,dimension(:) ,intent(in)     :: modelIndices   !! Indices of models whose named term should be evaluated
         type(StringArray) ,dimension(:) ,intent(in)     :: termNames     !! Name of evaluated term corresponding to each model 
+        logical           ,optional     ,intent(in)     :: accumulate    !! Optional flag to make the evaluator accumulate values instead of overriding them
 
     end subroutine initEvaluator
 !-----------------------------------------------------------------------------------------------------------------------------------