+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added slope limiter related unary transformations 
 - Variables can now be copied into/from arrays by passing a list of variable names to the container
 - Manipulators now called before first time step
+- The term evaluator manipulator can now accumulate values into the evaluation variable instead of overwriting it
 
 ### Bug Fixes